refactor(core): rename adb-hierarchy → maestro-hierarchy + compat shim
The Android view-hierarchy resolver is becoming the cross-platform Maestro
resolver (per percy-maestro/docs/plans/2026-04-27-001-feat-ios-element-regions-maestro-hierarchy-plan.md
Unit 1). Rename + shim is purely additive — no behavior change.

- Move src/adb-hierarchy.js → src/maestro-hierarchy.js (git mv preserves history).
- Move test/unit/adb-hierarchy.test.js → test/unit/maestro-hierarchy.test.js.
- Move test/fixtures/adb-hierarchy/ → test/fixtures/maestro-hierarchy/.
- Replace src/adb-hierarchy.js with a 5-line re-export shim. Removed in V1.1
  per the plan's deprecation guidance.
- Update api.js import to ./maestro-hierarchy.js.
- Update logger namespace from core:adb-hierarchy → core:maestro-hierarchy.
- Update file header to reflect cross-platform intent (the file body has been
  maestro-first for some time; the previous file name was always misleading).
- Update test describe block + import + fixture path.

Behavior unchanged. Subsequent units in Phase 1 will add the iOS branch and
api.js dispatch logic; this commit is just the rename so the diffs in those
units stay focused.
