+++ packages/core/test/fixtures/maestro-hierarchy/adversarial-trailer.txt
