+++ packages/core/test/fixtures/maestro-hierarchy/bad-bounds.xml
