+++ packages/core/test/fixtures/maestro-hierarchy/maestro-simple.json
