+++ packages/core/test/fixtures/maestro-hierarchy/simple.xml
