+++ packages/core/test/fixtures/maestro-hierarchy/with-trailer.txt
